Update projects : Move directory & Change name
