+++ app/src/main/java/com/gdsc_cau/vridge/data/dto/TtsDTO.kt
@@ -1,4 +1,4 @@
-package com.gdsc_cau.vridge.data.models
+package com.gdsc_cau.vridge.data.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
