+++ app/src/main/java/com/gdsc_cau/vridge/data/models/VoiceListResponse.kt
@@ -1,6 +1,6 @@
 package com.gdsc_cau.vridge.data.models
 
-data class VoiceListDTO(
+data class VoiceListResponse(
     var uid: String,
     var voiceList: List<String>
 )
